Trying to modify checkout and checkin
