+++ app/controllers/rentals_controller.rb
@@ -1,46 +1,54 @@
+require 'date'
+
 class RentalsController < ApplicationController
 
-  def checkout
-    # create a new rental with the movie_id and the customer_id
-    rental = Rental.new(rental_params)
+  # def checkout
+  #   # create a new rental with the movie_id and the customer_id
+  #   # rental = Rental.new(movie_id: movie_id, customer_id: customer_id, checkout_date: Date.today, due_date: Date.today + 7, currently_checked_out: true)
 
-    # save rental
-    if rental.save
-      render json: rental.as_json(only: [rental_params]), status: :ok
+  #   rental = Rental.new(rental_params)
+
+  #   puts "VVVVVVVVVVV"
+  #   puts "CHECKOUT METHOD"
+
+  #   # decrease movie inventory
+  #   movie = Movie.find(movie_id: movie_id)
+  #   movie.update(inventory: movie.inventory - 1)
+  #   # save rental
+  #   if rental.save
+  #     render json: rental.as_json(only: [rental_params]), status: :ok
+      
+  #   else
+  #     render json: { 
+  #       ok: false, 
+  #       errors: rental.errors.messages 
+  #     }, 
+  #     status: :bad_request
+  #   end
+  # end
+
+  # def checkin
+  #   rental = Rental.find_by(movie_id: rental_params[movie.id], customer_id: rental_params[customer_id])
+
+  #   movie.update(inventory: movie.inventory + 1)
+  #   rental.update(checkout_date: nil, due_date: nil, currently_checked_out: false)
     
-      # decrease movie inventory
-      movie.update(inventory: movie.inventory - 1)
-    else
-      render json: { 
-        ok: false, 
-        errors: rental.errors.messages 
-      }, 
-      status: :bad_request
-    end
-  end
+  #   if rental.save
+  #       render json: rental.as_json(only:[:movie_id, :customer_id]), status: :ok
+  #   else
+  #       render json: {
+  #           ok: false,
+  #           errors: rental.erros.messages
+  #       },
+  #       status: :bad_request
+  #   end
+  # end
 
-  def checkin
-    movie = Movie.find_by(id: rental_params[:id])
-    customer = Customer.find_by(id: rental_params[:id])
-    rental = Rental.new(movie_id: movie.id, customer_id: customer.id)
+  # private
 
-    if rental.save
-        movie.update(inventory: movie.inventory + 1) 
-        render json: rental.as_json(only:[rental_params]), status: :ok
-    else
-        render json: {
-            ok: false,
-            errors: rental.erros.messages
-        },
-        status: :bad_request
-    end
-  end
-
-  private
-
-  def rental_params
-      params.permit(:movie_id, :customer_id)
-  end
+  # def rental_params
+  #     params.permit(:movie_id, :customer_id)
+  # end
 end
